feat(Transaction, transactions): implemented logic to top up and privilege user

+++ app/models/admin/money_transfer.py
@@ -1,7 +1,9 @@
 from pydantic import BaseModel
 
+from decimal import Decimal
+
 
 class MoneyTransferRequest(BaseModel):
     sender_name: str
     recipient_name: str
-    amount: int
+    amount: Decimal
